requirements.txt
flask
google-adk[genai]
google-genai
